refactor(messaging): move consumer and producer setup from service to messaging module
