+++ llm-service/app/messaging/rabbitmq_consumer.py
